refactor : move ConversationManager/ConversationManagerInter to communication
